[CP-1090] Improve selection manager performance
