+++ packages/app/src/contacts/components/contact-group/contact-group-test-ids.enum.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export enum ContactGroupTestIdsEnum {
+  ContactListGroup = "contact-list-group",
+}
