+++ packages/app/src/contacts/components/contact-item/contact-item-test-ids.enum.ts
@@ -1,0 +1,9 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export enum ContactItemTestIdsEnum {
+  ContactRow = "contact-row",
+  ContactExportButton = "contact-export-button",
+}
