+++ packages/app/src/contacts/components/contact-list/contact-list-test-ids.enum.ts
@@ -4,10 +4,7 @@
  */
 
 export enum ContactListTestIdsEnum {
-  ContactRow = "contact-row",
   ContactListEmpty = "contact-list-empty",
   ContactListNoResult = "contact-list-no-result",
   ContactListLoading = "contact-list-loading",
-  ContactListGroup = "contact-list-group",
-  ContactExportButton = "contact-export-button",
 }
