+++ packages/app/src/contacts/components/contact-simple-list-item-placeholder/contact-simple-list-item-placeholder.component.tsx
@@ -9,7 +9,7 @@ import {
   Col,
   TextPlaceholder,
 } from "App/__deprecated__/renderer/components/core/table/table.component"
-import { AvatarPlaceholder } from "App/contacts/components/contact-list/contact-list.styled"
+import { AvatarPlaceholder } from "App/contacts/components/contact-item/contact-item.styled"
 import { createFullName } from "App/contacts/helpers/contacts.helpers"
 import { ContactSimpleListItemPlaceholderProps } from "App/contacts/components/contact-simple-list-item-placeholder/contact-simple-list-item-placeholder.interface"
 import { ContactSimpleListItemPlaceholderTestIdsEnum } from "App/contacts/components/contact-simple-list-item-placeholder/contact-simple-list-item-placeholder-test-ids.enum"
